refactor: extract handlers from main.rs and increase test coverage from 34% to 40%
- Extract all handler functions from main.rs to new handlers.rs module
- Reduce main.rs from 603 lines to 61 lines (90% reduction)
- Add 37 comprehensive test cases across 3 new test files:
  - tests/handlers_tests.rs (16 tests for synchronous handlers)
  - tests/handlers_unit_tests.rs (11 tests for edge cases)
  - tests/handlers_async_tests.rs (10 tests for async handlers)
- Expand tests/proxy_tests.rs with additional process spawning tests
- Improve coverage: handlers.rs 101/293 (34.5%), proxy.rs 33/98 (33.7%)

Coverage improved from 34.15% (224/656 lines) to 40.14% (297/740 lines)
This represents a +5.99% absolute increase and +17.5% relative improvement.

The refactoring separates business logic from CLI entry point, making
the codebase significantly more maintainable and testable.
