+++ src/lib.rs
@@ -2,5 +2,6 @@ pub mod auth;
 pub mod cli;
 pub mod config;
 pub mod filters;
+pub mod handlers;
 pub mod keyring_token_store;
 pub mod proxy;
